add validation tests, replaced xs and ys, fix cpu_basic and gpu_basic

+++ src/utils/helpers.hpp
@@ -1,5 +1,5 @@
 #pragma once
 
-#define SECOND(x) (x / 1000000000)
-#define MILLISECOND(x) (x / 1000000)
-#define MICROSECOND(x) (x / 1000)
+#define SECOND(x) ((float)x / 1000000000)
+#define MILLISECOND(x) ((float)x / 1000000)
+#define MICROSECOND(x) ((float)x / 1000)
